moved open_data to its own private file module; moved ASReviewData to its own private file module; made _get_filename_from_url a class method of ASReviewData; updated imports and usage throughout
