+++ asreview/__init__.py
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from asreview.data.base import ASReviewData
-from asreview.data.base import load_data
 from asreview.io.utils import list_readers
 from asreview.io.utils import list_writers
 from asreview.models.feature_extraction.embedding_lstm import load_embedding
@@ -23,6 +21,8 @@ from asreview.project import ASReviewProject
 from asreview.project import open_state
 from asreview.utils import asreview_path
 from asreview.utils import get_data_home
+from .data import ASReviewData
+from .data import load_data
 
 from ._version import get_versions
